safari-534.53-branch/LayoutTests/fast/canvas/canvas-state-intact-after-putImageData-expected.txt

Test that the rendering context state is intact after a call to putImageData()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking initial state for sanity
PASS ctx.fillStyle is '#ff0000'
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
PASS imgdata[4] is 0
PASS imgdata[5] is 0
PASS imgdata[6] is 0
PASS imgdata[7] is 0
Calling putImageData()
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
Checking if state is intact
PASS ctx.fillStyle is '#ff0000'
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255
PASS successfullyParsed is true

TEST COMPLETE
